requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.7
billiard==3.3.0.23
celery==3.1.26.post2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-celery==3.3.1
django-crispy-forms==1.9.0
django-formtools==2.2
django-import-export==2.0.2
django-redis==4.11.0
django-reversion==3.0.7
djangorestframework==3.11.0
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
jdcal==1.4.1
kombu==3.0.37
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
pytz==2019.3
PyYAML==5.3.1
redis==2.10.6
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
